pyproject.toml
[tool.poetry]
name = "bs-ui"
version = "0.1.0"
description = ""
authors = ["Christopher Michael Sutton <christopher.sutton12@gmail.com>"]

[tool.poetry.scripts]
start = "bs_ui.main:start"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
PySide6 = "^6.2.1"
matplotlib = "^3.5.0"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
